+++ R/plot_trait_distribution_beeswarm.R
@@ -1,4 +1,4 @@
-#' @title Trait distribution
+#' @title Beeswarm Trait distribution 
 #' @description Plots distribution of trait values by a  grouping variable using ggbeeswarm package  
 #'
 #' @param austraits austraits data object
@@ -15,9 +15,11 @@
 #' }
 #' @author Daniel Falster - daniel.falster@unsw.edu.au
 #' @export
-#' @importFrom rlang .data
+
 #
 plot_trait_distribution_beeswarm <- function(austraits, plant_trait_name, y_axis_category, highlight=NA, hide_ids = FALSE) {
+  # Determine version
+  version <- austraits$build_info$version %>% as.character()
   
   # Subset data to this trait
   austraits_trait <- extract_trait(austraits, plant_trait_name)
@@ -33,16 +35,16 @@ plot_trait_distribution_beeswarm <- function(austraits, plant_trait_name, y_axis
     factor(p, levels=names(my_shapes))
   }
   
-  tax_info  <- austraits_trait$taxa %>% dplyr::select(.data$taxon_name, .data$family)
+  tax_info  <- austraits_trait$taxa %>% dplyr::select(taxon_name, family)
   
   data <- 
     austraits_trait$traits %>%
-    dplyr::mutate(shapes = as_shape(.data$value_type)) %>%
+    dplyr::mutate(shapes = as_shape(value_type)) %>%
     dplyr::left_join(by = "taxon_name", tax_info)
   
   # Define grouping variables and derivatives
   if(!y_axis_category %in% names(data)){
-    stop("incorrect grouping variable")
+    stop("Incorrect grouping variable! Currently implemented for `family` or `dataset_id`")
   }
   
   # define grouping variable, ordered by group-level by mean values
@@ -62,11 +64,19 @@ plot_trait_distribution_beeswarm <- function(austraits, plant_trait_name, y_axis
   
   # set colour of group to highlight
   if(!is.na(highlight) & highlight %in% data$Group) {
-    data <- dplyr::mutate(data, colour = ifelse(.data$Group %in% highlight, "c", .data$colour))
+    data <- dplyr::mutate(data, colour = ifelse(Group %in% highlight, "c", colour))
   }
   
   # Check range on x-axis
-  vals <- austraits_trait$definitions$traits$elements[[plant_trait_name]]$values
+  if(package_version(version) <= '3.0.2'){
+  vals <- austraits_trait$definitions$traits$elements[[plant_trait_name]]$value
+  }
+  
+  if(package_version(version) > '3.0.2'){
+  vals <- list(minimum = purrr::pluck(austraits_trait, "definitions", plant_trait_name, "allowed_values_min"),
+           maximum = purrr::pluck(austraits_trait, "definitions", plant_trait_name, "allowed_values_max"))
+  }
+  
   range <- (vals$maximum/vals$minimum)
   
   # Check range on y-axis
@@ -75,8 +85,8 @@ plot_trait_distribution_beeswarm <- function(austraits, plant_trait_name, y_axis
   
   # Top plot - plain histogram of data
   p1 <-
-    ggplot2::ggplot(data, ggplot2::aes(x=.data$value)) +
-    ggplot2::geom_histogram(ggplot2::aes(y = .data$..density..), color="darkgrey", fill="darkgrey", bins=50) +
+    ggplot2::ggplot(data, ggplot2::aes(x=value)) +
+    ggplot2::geom_histogram(ggplot2::aes(y = ..density..), color="darkgrey", fill="darkgrey", bins=50) +
     ggplot2::geom_density(color="black") +
     ggplot2::xlab("") + ggplot2::ylab("All data") +
     ggplot2::theme_bw()  +
@@ -89,7 +99,7 @@ plot_trait_distribution_beeswarm <- function(austraits, plant_trait_name, y_axis
     )
   # Second plot -- dots by groups, using ggbeeswarm package
   p2 <-
-    ggplot2::ggplot(data, ggplot2::aes(x = .data$value, y = .data$Group, colour = .data$colour, shape = .data$shapes)) +
+    ggplot2::ggplot(data, ggplot2::aes(x = value, y = Group, colour = colour, shape = shapes)) +
     ggbeeswarm::geom_quasirandom(groupOnX=FALSE) +
     ggplot2::ylab(paste("By ", y_axis_category)) +
     # inclusion of custom shapes: for min, mean, unknown