+++ NEWS.md
@@ -1,7 +1,24 @@
+# austraits 2.1.1
 
-# austraits 1.1.1
-
-* Improved vignette building process
-* Improved extract_ functions
-* Rebuilt website
-
+-   Minor bug fix in `extract_trait`
+-   Minor bug fix in `as_wide_table`
+-   `site` & `site_name` renamed to `location` & `location_name` within functions
+-   `austraits$sites` renamed to `austraits$locations`
+-   `plot_site_locations` is deprecated, use `plot_locations`
+-   `join_sites` is deprecated, use `join_locations`
+-   The following functions have been updated to work with AusTraits > v3.0.2, as well <=3.0.2
+    - `join_taxonomy`
+    - `join_locations`
+    - `join_contexts`
+    - `join_methods`
+    - `join_all`
+    - `trait_pivot_wider`
+    - `trait_pivot_longer`
+    - `as_wide_table`
+    - `extract_taxa`
+    - `extract_trait`
+    - `extract_dataset`
+    - `plot_locations`
+    - `plot_trait_distribution_beeswarm`
+    - `print.austraits`
+